Fixed some bugs and performances issues, also resolved some compiler warnings

+++ DV1573---UD1448/Renderer/Renderer.cpp
@@ -205,7 +205,7 @@ void Renderer::render() {
 		for (GameObject* object : m_staticObjects)
 		{
 			//Then through all of the meshes
-			for (size_t j = 0; j < object->getMeshesCount(); j++)
+			for (int j = 0; j < object->getMeshesCount(); j++)
 			{
 				modelMatrix = glm::mat4(1.0f);
 				//Fetch the current mesh and its transform
@@ -230,7 +230,7 @@ void Renderer::render() {
 		for (GameObject* object : m_anistaticObjects)
 		{
 			//Then through all of the meshes
-			for (size_t j = 0; j < object->getMeshesCount(); j++)
+			for (int j = 0; j < object->getMeshesCount(); j++)
 			{
 				modelMatrix = glm::mat4(1.0f);
 				//Fetch the current mesh and its transform
@@ -303,7 +303,7 @@ void Renderer::render() {
 	for (GameObject* object : m_staticObjects)
 	{
 		//Then through all of the meshes
-		for (size_t j = 0; j < object->getMeshesCount(); j++)
+		for (int j = 0; j < object->getMeshesCount(); j++)
 		{
 			//Fetch the current mesh and its transform
 			mesh = MeshMap::getInstance()->getMesh(object->getMeshName(j));
@@ -334,7 +334,7 @@ void Renderer::render() {
 			}
 			
 			//Then through all of the meshes
-			for (size_t j = 0; j < object->getMeshesCount(); j++)
+			for (int j = 0; j < object->getMeshesCount(); j++)
 			{
 				//Fetch the current mesh and its transform
 				mesh = MeshMap::getInstance()->getMesh(object->getMeshName(j));
@@ -377,7 +377,7 @@ void Renderer::render() {
 	for (GameObject* object : m_anistaticObjects)
 	{
 		//Then through all of the meshes
-		for (size_t j = 0; j < object->getMeshesCount(); j++)
+		for (int j = 0; j < object->getMeshesCount(); j++)
 		{
 			//Fetch the current mesh and its transform
 			mesh = MeshMap::getInstance()->getMesh(object->getMeshName(j));
